♻️ Refactor: TransparentButton and PlayHeader components
