+++ client/src/features/artist/components/Artist.tsx
@@ -6,9 +6,9 @@ import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
 import { useEffect } from 'react';
 import { getArtist } from '../artistThunks.ts';
 import LoadingScreen from '../../../ui/StatusScreens/LoadingScreen.tsx';
-import styles from '../../../components/PlayHeader/PlayHeader.module.scss';
-import PlayButton from '../../../components/PlayHeader/PlayButton.tsx';
-import TransparentButton from '../../../components/PlayHeader/TransparentButton.tsx';
+import styles from '../../../ui/PlayHeader/PlayHeader.module.scss';
+import PlayButton from '../../../ui/PlayHeader/PlayButton.tsx';
+import TransparentButton from '../../../ui/Button/TransparentButton.tsx';
 import { followUser, unfollowUser } from '../../auth/userThunks.ts';
 import UploadSongDialog from './UploadSongDialog.tsx';
 import NotFound from '../../../ui/StatusScreens/NotFound.tsx';
