+++ client/src/features/artist/components/UploadSongDialog.tsx
@@ -2,9 +2,9 @@ import styles from './UploadSongDialog.module.scss';
 import Dialog from '../../../ui/Dialog/Dialog.tsx';
 import DialogContent from '../../../ui/Dialog/DialogContent.tsx';
 import DialogTrigger from '../../../ui/Dialog/DialogTrigger.tsx';
-import TransparentButton from '../../../components/PlayHeader/TransparentButton.tsx';
+import TransparentButton from '../../../ui/Button/TransparentButton.tsx';
 import Input from '../../../ui/Input/Input.tsx';
-import Button from '../../../ui/Button/Button';
+import Button from '../../../ui/Button/Button.tsx';
 import { FormEvent, useRef } from 'react';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
 import { uploadSong } from '../artistThunks.ts';
