+++ client/src/features/playlist/components/PlaylistActions.module.scss
@@ -1,0 +1,11 @@
+.actionsButton {
+  @apply size-6 text-neutral-400;
+
+  &:hover {
+    @apply text-white scale-105;
+  }
+
+  &:active {
+    @apply text-neutral-500 scale-100;
+  }
+}
