+++ client/src/layout/Nav/components/UserActions.jsx
@@ -1,4 +1,4 @@
-import Button from '../../../ui/Button/Button.jsx';
+import Button from '../../../ui/Button/Button.tsx';
 import NavTooltip from './NavTooltip.jsx';
 import NavButton from './NavButton.jsx';
 import { RiNotification3Line } from 'react-icons/ri';
