+++ client/src/ui/Button/Button.tsx
@@ -1,6 +1,11 @@
 import styles from './Button.module.scss';
+import { ReactNode } from 'react';
 
-const Button = ({ children }) => {
+interface ButtonProps {
+  children: ReactNode;
+}
+
+const Button = ({ children }: ButtonProps) => {
   return <button className={styles.button}>{children}</button>;
 };
