+++ client/src/ui/ImageHeader/ImageHeader.module.scss
@@ -1,7 +1,7 @@
 .header {
   @apply p-5;
-  @apply grid grid-cols-[auto_1fr] gap-5 items-end
-    
+  @apply grid grid-cols-[auto_1fr] gap-5 items-end;
+
   height: 300px;
 }
