+++ client/src/ui/PlayHeader/PlayButton.module.scss
@@ -1,0 +1,12 @@
+.playButton {
+  @apply size-14 p-4;
+  @apply text-black bg-green-500 rounded-full cursor-pointer;
+
+  &:hover {
+    @apply bg-green-400 scale-105;
+  }
+
+  &:active {
+    @apply bg-green-600 scale-100;
+  }
+}
