+++ client/src/ui/PlayHeader/PlayButton.tsx
@@ -1,5 +1,5 @@
 import { RiPlayLargeFill } from 'react-icons/ri';
-import styles from './PlayHeader.module.scss';
+import styles from './PlayButton.module.scss';
 
 const PlayButton = () => {
   return <RiPlayLargeFill className={styles.playButton} role="button" />;
