+++ client/src/ui/PlayHeader/PlayHeader.module.scss
@@ -1,0 +1,3 @@
+.playerHeader {
+  @apply p-5 flex items-center gap-6;
+}
